requirements.txt
speechrecognition 
google-generativeai 
pyttsx3 
pyaudio
